backend/requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Pydantic for data validation
pydantic==2.5.0

# CORS middleware and security
python-multipart==0.0.6
slowapi==0.1.9
redis==5.0.1

# AI and ML dependencies
openai==1.3.0
pinecone-client==2.2.4

# HTTP client for external API calls
httpx==0.25.2
aiohttp==3.9.1

# Environment variable management
python-dotenv==1.0.0

# Data processing
numpy==1.24.3
pandas==2.1.4

# Async support
asyncio==3.4.3

# Logging and monitoring
structlog==23.2.0

# Testing (optional, for development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# System monitoring
psutil==5.9.6

# Development tools (optional)
black==23.11.0
flake8==6.1.0
mypy==1.7.1
